Refactor out some global space to allow for testable code and handling error conditions before flask is completely initialized.
